requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Google APIs
google-generativeai==0.8.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0

# Data Processing
pandas==2.0.3
numpy==1.24.3

# HTTP & API
requests==2.31.0
httpx==0.25.0

# Environment
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
